Initial Test Version
TEST_MIT.ino is a program for the Photon that exposes two cloud
functions: one to turn an LED on and off, and another to move a servo.
The LED and servo pins are compatible with the WLD so that WLD hardware
can be used for testing.

ParticleTest.aia is an MIT app inventor 2 source file that creates a
basic app for turning the LED on and of and for moving the servo.